Simplify error message matching in unit tests

+++ geometry/render_gltf_client/test/internal_render_client_test.cc
@@ -28,10 +28,12 @@ namespace internal {
 namespace fs = drake::filesystem;
 
 using Params = RenderEngineGltfClientParams;
+using geometry::render::ClippingRange;
 using geometry::render::ColorRenderCamera;
 using geometry::render::DepthRange;
 using geometry::render::DepthRenderCamera;
 using geometry::render::RenderCameraCore;
+using systems::sensors::CameraInfo;
 using systems::sensors::ImageDepth32F;
 using systems::sensors::ImageLabel16I;
 using systems::sensors::ImageRgba8U;
@@ -71,56 +73,43 @@ class RenderClientTest : public ::testing::Test {
 
 // Constructor / destructor ----------------------------------------------------
 TEST_F(RenderClientTest, Constructor) {
-  const std::string base_url{"127.0.0.1:8000"};
-  const std::string render_endpoint{"render"};
-  const bool verbose = false;
+  // Verify default attributes are set properly given a default params.
+  const RenderClient default_client{Params{}};
+  EXPECT_EQ(
+      default_client.get_params().GetUrl(), "http://127.0.0.1:8000/render");
+  EXPECT_EQ(default_client.get_params().verbose, false);
+  EXPECT_EQ(default_client.get_params().no_cleanup, false);
 
-  // Verify attributes after valid construction.
-  auto make_client_and_verify = [&](bool p_no_cleanup) {
-    std::string temp_directory;
-    {
-      RenderClient client{Params{base_url, render_endpoint, std::nullopt,
-                                 verbose, p_no_cleanup}};
-      temp_directory = client.temp_directory();
-      EXPECT_EQ(client.get_params().GetUrl(), "127.0.0.1:8000/render");
-      EXPECT_EQ(client.get_params().verbose, verbose);
-      EXPECT_EQ(client.get_params().no_cleanup, p_no_cleanup);
-      EXPECT_TRUE(fs::is_directory(client.temp_directory()));
-    }  // Client is deleted.
-    if (p_no_cleanup) {
-      EXPECT_TRUE(fs::is_directory(temp_directory));
-    } else {
-      EXPECT_FALSE(fs::is_directory(temp_directory));
-    }
-  };
-  make_client_and_verify(true);
-  make_client_and_verify(false);
+  // Verify attributes are set properly given a specific params.
+  const std::string base_url{"http://127.0.0.1:1234"};
+  const std::string render_endpoint{"testing"};
+  const bool verbose = true;
+  const bool no_cleanup = true;
+  const RenderClient client{
+      Params{base_url, render_endpoint, std::nullopt, verbose, no_cleanup}};
+  EXPECT_EQ(client.get_params().GetUrl(),
+            base_url + "/" + render_endpoint);
+  EXPECT_EQ(client.get_params().verbose, verbose);
+  EXPECT_EQ(client.get_params().no_cleanup, no_cleanup);
 }
 
 TEST_F(RenderClientTest, Destructor) {
-  const std::string base_url{"127.0.0.1:8000"};
-  const std::string render_endpoint = "render";
-  const bool verbose = false;
+  for (const bool no_cleanup : {true, false}) {
+    std::string temp_dir_path;
+    {
+      const RenderClient client{Params{.no_cleanup = no_cleanup}};
+      temp_dir_path = client.temp_directory();
+      EXPECT_TRUE(fs::is_directory(temp_dir_path));
+    }  // Client is deleted.
 
-  std::string temp_dir_path;
-  // Construction with no_cleanup=false: temp_directory should be gone.
-  {
-    const RenderClient client{
-        Params{base_url, render_endpoint, std::nullopt, verbose, false}};
-    temp_dir_path = client.temp_directory();
-    EXPECT_TRUE(fs::is_directory(temp_dir_path));
-  }  // Client is deleted.
-  EXPECT_FALSE(fs::is_directory(temp_dir_path));
-
-  // Construction with no_cleanup=true: temp_directory should remain.
-  {
-    const RenderClient client{
-        Params{base_url, render_endpoint, std::nullopt, verbose, true}};
-    temp_dir_path = client.temp_directory();
-    EXPECT_TRUE(fs::is_directory(temp_dir_path));
-  }  // Client is deleted.
-  EXPECT_TRUE(fs::is_directory(temp_dir_path));
-  fs::remove(temp_dir_path);
+    // Test whether temp_directory is handled properly.
+    if (no_cleanup) {
+      EXPECT_TRUE(fs::is_directory(temp_dir_path));
+      fs::remove(temp_dir_path);
+    } else {
+      EXPECT_FALSE(fs::is_directory(temp_dir_path));
+    }
+  }
 }
 
 // RenderOnServer --------------------------------------------------------------
@@ -160,8 +149,7 @@ class FieldCheckService : public HttpService {
                     const std::string& scene_sha256,
                     const std::optional<std::string>& mime_type = std::nullopt,
                     const std::optional<DepthRange>& depth_range = std::nullopt)
-      : HttpService(),
-        camera_core_{camera_core},
+      : camera_core_{camera_core},
         image_type_{image_type},
         scene_path_{scene_path},
         scene_sha256_{scene_sha256},
@@ -186,10 +174,10 @@ class FieldCheckService : public HttpService {
     } else {  // image_type_ := RenderImageType::kLabel16I
       EXPECT_EQ(data_fields.at("image_type"), "label");
     }
-    const auto& intrinsics = camera_core_.intrinsics();
+    const CameraInfo& intrinsics = camera_core_.intrinsics();
     EXPECT_EQ(data_fields.at("width"), std::to_string(intrinsics.width()));
     EXPECT_EQ(data_fields.at("height"), std::to_string(intrinsics.height()));
-    const auto& clipping = camera_core_.clipping();
+    const ClippingRange& clipping = camera_core_.clipping();
     EXPECT_EQ(data_fields.at("near"), std::to_string(clipping.near()));
     EXPECT_EQ(data_fields.at("far"), std::to_string(clipping.far()));
     EXPECT_EQ(data_fields.at("focal_x"), std::to_string(intrinsics.focal_x()));
@@ -202,7 +190,7 @@ class FieldCheckService : public HttpService {
               std::to_string(intrinsics.center_y()));
     // min_depth and max_depth should only be provided for depth renders.
     if (image_type_ == RenderImageType::kDepthDepth32F) {
-      const auto& range = depth_range_.value();
+      const DepthRange& range = depth_range_.value();
       EXPECT_EQ(data_fields.at("min_depth"), std::to_string(range.min_depth()));
       EXPECT_EQ(data_fields.at("max_depth"), std::to_string(range.max_depth()));
     } else {
@@ -260,7 +248,7 @@ using PostFormCallback = typename std::function<HttpResponse()>;
 class ProxyService : public HttpService {
  public:
   explicit ProxyService(const PostFormCallback& callback)
-      : HttpService(), post_form_callback_{callback} {}
+      : post_form_callback_{callback} {}
 
   HttpResponse DoPostForm(const std::string& /* temp_directory */,
                           const std::string& /* url */,
@@ -294,7 +282,6 @@ TEST_F(RenderClientTest, RenderOnServer) {
       Params{base_url, render_endpoint, std::nullopt, verbose, no_cleanup});
   // All the files generated under `temp_dir_path` will be purged in the end.
   const auto temp_dir_path = fs::path(client.temp_directory());
-  const std::string url = client.get_params().GetUrl();
 
   // Create a fake scene to "upload" to the "server" and fake response file.
   const auto fake_scene_path = (temp_dir_path / "fake_scene.gltf").string();
@@ -337,13 +324,8 @@ TEST_F(RenderClientTest, RenderOnServer) {
      use DRAKE_EXPECTS_THROWS_MESSAGE all looking for the same error being
      raised, as no image response is provided.  However, the bulk of the test
      takes place in the assertions in FieldCheckService::PostForm. */
-    const auto expected_message = fmt::format(
-        "\\s*ERROR doing POST:\\s*"        // ERROR doing POST:
-        "URL:\\s*{}\\s*"                   // URL:              {url}
-        "Service Message:\\s*None\\.\\s*"  // Service Message:  None.
-        "HTTP Code:\\s*500\\s*"            // Http Code:        {code}
-        "Server Message:\\s*None\\.\\s*",  // Server Message:   None.
-        url);
+    const std::string expected_throw_pattern =
+        "[\\s\\S]*ERROR doing POST:[\\s\\S]*HTTP Code:[\\s\\S]*500[\\s\\S]*";
 
     // Check fields for a color render.
     client.SetHttpService(std::make_unique<FieldCheckService>(
@@ -352,7 +334,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(),
                               RenderImageType::kColorRgba8U, fake_scene_path),
-        expected_message);
+        expected_throw_pattern);
 
     /* Check fields for a depth render.  This test also includes a verification
      that the provided mime_type is propagated correctly.  There is no special
@@ -365,7 +347,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
         client.RenderOnServer(depth_camera.core(),
                               RenderImageType::kDepthDepth32F, fake_scene_path,
                               "test/mime_type", depth_camera.depth_range()),
-        expected_message);
+        expected_throw_pattern);
 
     // Check fields for a label render.
     client.SetHttpService(std::make_unique<FieldCheckService>(
@@ -374,7 +356,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(), RenderImageType::kLabel16I,
                               fake_scene_path),
-        expected_message);
+        expected_throw_pattern);
   }
 
   {
@@ -384,13 +366,8 @@ TEST_F(RenderClientTest, RenderOnServer) {
      all produce 'None.' as the server message.
      NOTE: file extensions do not matter. */
     // NOTE: all tests using this must use http code 400.
-    const auto message_template = fmt::format(
-        "\\s*ERROR doing POST:\\s*"        // ERROR doing POST:
-        "URL:\\s*{}\\s*"                   // URL:              {url}
-        "Service Message:\\s*None\\.\\s*"  // Service Message:  None.
-        "HTTP Code:\\s*400\\s*"            // Http Code:        {code}
-        "Server Message:\\s*{{}}\\s*",     // Server Message:   {message}
-        url);
+    const std::string expected_throw_pattern =
+        "[\\s\\S]*ERROR doing POST:[\\s\\S]*Server Message:\\s*{}[\\s\\S]*";
 
     PostFormCallback callback{nullptr};
     // Case 1: edge case, service populated data_path but file is length 0.
@@ -405,7 +382,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(),
                               RenderImageType::kColorRgba8U, fake_scene_path),
-        fmt::format(message_template, "None\\."));
+        fmt::format(expected_throw_pattern, "None\\."));
 
     // Case 2: edge case, bad response but provided message "too long".
     callback = [&]() {
@@ -420,7 +397,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(), RenderImageType::kLabel16I,
                               fake_scene_path),
-        fmt::format(message_template, "None\\."));
+        fmt::format(expected_throw_pattern, "None\\."));
 
     // Case 3: edge case, message provided of valid length but cannot be opened.
     const auto bad_permission_path =
@@ -437,8 +414,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(),
                               RenderImageType::kColorRgba8U, fake_scene_path),
-        fmt::format("RenderOnServer: cannot open file {}",
-                    bad_permission_path));
+        fmt::format(expected_throw_pattern, "None\\."));
 
     // Case 4: server response that can be read.
     const auto response_text = "You are not a valid request :p";
@@ -453,7 +429,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(), RenderImageType::kLabel16I,
                               fake_scene_path),
-        fmt::format(message_template, response_text));
+        fmt::format(expected_throw_pattern, response_text));
   }
 
   {
@@ -463,10 +439,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(),
                               RenderImageType::kColorRgba8U, fake_scene_path),
-        fmt::format(
-            "ERROR doing POST to {}, HTTP code=200: the server was supposed to"
-            " respond with a file but did not.",
-            url));
+        "ERROR doing POST.*supposed to respond with a file but did not.");
   }
 
   {
@@ -480,10 +453,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(), RenderImageType::kLabel16I,
                               fake_scene_path),
-        fmt::format(
-            "ERROR doing POST to {}, HTTP code=200: the service responded with "
-            "a file path '{}' but the file does not exist.",
-            url, response_path));
+        "ERROR doing POST.*the file does not exist.");
   }
 
   {
@@ -500,12 +470,7 @@ TEST_F(RenderClientTest, RenderOnServer) {
     DRAKE_EXPECT_THROWS_MESSAGE(
         client.RenderOnServer(color_camera.core(),
                               RenderImageType::kColorRgba8U, fake_scene_path),
-        fmt::format(
-            "RenderClient: while trying to render the scene '{}' with a sha256 "
-            "hash of '{}', the file returned by the server saved in '{}' is "
-            "not understood as an image type that is supported, i.e., PNG or "
-            "TIFF.",
-            fake_scene_path, fake_scene_sha256, response_path));
+        ".*is not understood as an image type that is supported.*");
   }
 
   {
